refactor: centralize receipts and reports-filter logic in packages/shared
Addresses the remaining review findings from the previous pass:

F4 — ALLOWED_STATUSES / ALLOWED_CURRENCIES Set constructions duplicated
across three routes. Adds ORDER_STATUS_SET and POPULAR_CURRENCIES_SET
(ReadonlySet<string>) to packages/shared alongside the existing list
exports. All three routes now read the canonical Set.

F5 — buildOrderFilters + helpers duplicated between admin's orders and
export routes. Extracts to a new _shared/reportsFilters.ts with one
buildAdminOrderFilters(searchParams). Both routes shrink ~80 lines.
Incidentally fixes a pre-existing smell in the orders route's
addDateFilters that reassigned its parameters after validation.

F6 / F7 / F8 — receipts bucket name, signed-URL TTL, safe-path regex,
and env-reading inconsistency in admin's receiptSignedUrls.ts:

  - Adds packages/shared/src/constants/receipts.ts with RECEIPTS_BUCKET
    and RECEIPT_URL_TTL_SECONDS.
  - Adds packages/shared/src/utils/receiptPath.ts with SAFE_RECEIPT_PATH,
    toSafeReceiptPath (null-safe) and assertSafeReceiptPath (throwing) —
    one regex used by both apps so admin and payments cannot drift.
  - Payments domain/constants.ts now re-exports the shared bucket+TTL.
  - Payments domain/receipt.ts delegates toSafeStoragePath /
    assertSafeStoragePath to the shared assertSafeReceiptPath.
  - Admin's receiptSignedUrls.ts reads SERVICE_ROLE_KEY at module load
    (matches the other URL constants) and drops the local
    SECONDS_PER_HOUR + SAFE_RECEIPT_PATH duplicates.

Net diff: −52 lines despite adding 4 new files. Behavior unchanged —
all 1034 unit tests still pass; semantics of toSafeStoragePath (throws)
and toSafeReceiptPath (returns null) preserved on the per-app boundary.
