+++ apps/admin/src/app/api/admin/reports/orders/route.ts
@@ -1,7 +1,5 @@
 /* eslint-disable i18next/no-literal-string */
 import { NextResponse } from "next/server";
-import { ORDER_STATUS_LIST } from "shared/constants/orders";
-import { POPULAR_CURRENCIES } from "shared/utils/currencies";
 
 import {
   adminFetch,
@@ -11,13 +9,10 @@ import {
   INTERNAL_SERVER_ERROR_STATUS,
 } from "@/app/api/admin/_shared/adminRest";
 import { signReceiptPath } from "@/app/api/admin/_shared/receiptSignedUrls";
-
-const ALLOWED_STATUSES = new Set<string>(ORDER_STATUS_LIST);
-const ALLOWED_CURRENCIES = new Set<string>(POPULAR_CURRENCIES);
+import { buildAdminOrderFilters } from "@/app/api/admin/_shared/reportsFilters";
 
 const ADMIN_REPORTS = "admin.reports";
 const MAX_LIMIT = 10_000;
-const ISO_DATE_LENGTH = 10;
 const ORDERS_SELECT =
   "id,created_at,payment_status,total,currency,transfer_number,receipt_url,user_id,seller_id";
 const ITEMS_SELECT =
@@ -49,97 +44,6 @@ interface UserProfileRow {
   id: string;
   email: string;
   display_name: string | null;
-}
-
-const ISO_DATE_REGEX = /^\d{4}-\d{2}-\d{2}$/;
-
-function isValidIsoDate(value: string): boolean {
-  return ISO_DATE_REGEX.test(value) && !Number.isNaN(Date.parse(value));
-}
-
-function isValidAmount(value: string): boolean {
-  const num = Number.parseFloat(value);
-  return !Number.isNaN(num) && num >= 0;
-}
-
-function addDateFilters(
-  filters: Record<string, string>,
-  dateFrom: string | null,
-  dateTo: string | null,
-): void {
-  const validFrom = dateFrom && isValidIsoDate(dateFrom) ? dateFrom : null;
-  const validTo = dateTo && isValidIsoDate(dateTo) ? dateTo : null;
-  dateFrom = validFrom;
-  dateTo = validTo;
-  if (dateFrom && dateTo) {
-    const end = new Date(dateTo);
-    end.setDate(end.getDate() + 1);
-    filters["created_at"] = `gte.${dateFrom}`;
-    filters["and"] =
-      `(created_at.lt.${end.toISOString().slice(0, ISO_DATE_LENGTH)})`;
-  } else if (dateFrom) {
-    filters["created_at"] = `gte.${dateFrom}`;
-  } else if (dateTo) {
-    const end = new Date(dateTo);
-    end.setDate(end.getDate() + 1);
-    filters["created_at"] = `lt.${end.toISOString().slice(0, ISO_DATE_LENGTH)}`;
-  }
-}
-
-function addAmountFilters(
-  filters: Record<string, string>,
-  amountMin: string | null,
-  amountMax: string | null,
-): void {
-  const validMin = amountMin && isValidAmount(amountMin) ? amountMin : null;
-  const validMax = amountMax && isValidAmount(amountMax) ? amountMax : null;
-  if (validMin && validMax) {
-    filters["total"] = `gte.${validMin}`;
-    const existing = filters["and"] ?? "";
-    filters["and"] = existing
-      ? `${existing},(total.lte.${validMax})`
-      : `(total.lte.${validMax})`;
-  } else if (validMin) {
-    filters["total"] = `gte.${validMin}`;
-  } else if (validMax) {
-    filters["total"] = `lte.${validMax}`;
-  }
-}
-
-function buildOrderFilters(
-  searchParams: URLSearchParams,
-): Record<string, string> {
-  const filters: Record<string, string> = {};
-
-  addDateFilters(
-    filters,
-    searchParams.get("dateFrom"),
-    searchParams.get("dateTo"),
-  );
-  addAmountFilters(
-    filters,
-    searchParams.get("amountMin"),
-    searchParams.get("amountMax"),
-  );
-
-  const status = searchParams.get("status");
-  const sellerId = searchParams.get("sellerId");
-  const buyerId = searchParams.get("buyerId");
-  const currency = searchParams.get("currency");
-
-  if (status && ALLOWED_STATUSES.has(status)) {
-    filters["payment_status"] = `eq.${status}`;
-  }
-  if (sellerId) filters["seller_id"] = `eq.${sellerId}`;
-  if (buyerId) filters["user_id"] = `eq.${buyerId}`;
-  if (currency) {
-    const normalized = currency.toUpperCase();
-    if (ALLOWED_CURRENCIES.has(normalized)) {
-      filters["currency"] = `eq.${normalized}`;
-    }
-  }
-
-  return filters;
 }
 
 async function fetchOrderItems(
@@ -179,7 +83,7 @@ export async function GET(request: Request) {
 
   try {
     const { searchParams } = new URL(request.url);
-    const filters = buildOrderFilters(searchParams);
+    const filters = buildAdminOrderFilters(searchParams);
     const productId = searchParams.get("productId");
 
     const ordersResponse = await adminFetch(
