+++ apps/payments/src/shared/domain/receipt.ts
@@ -1,3 +1,5 @@
+import { assertSafeReceiptPath } from "shared/utils/receiptPath";
+
 import {
   ACCEPTED_RECEIPT_MIME_TYPES,
   MAX_RECEIPT_SIZE_BYTES,
@@ -9,11 +11,6 @@ const MAX_RECEIPT_FILENAME_LENGTH = 64;
 
 // Only alphanumeric characters and hyphens — no slashes, dots, or other path chars.
 const SAFE_STORAGE_SEGMENT = /^([a-zA-Z0-9_-]+)$/;
-
-// Whitelist for full storage paths: exactly one slash, both segments alphanumeric+hyphen+underscore,
-// second segment has a known image extension. Capture groups used to reconstruct below.
-const SAFE_RECEIPT_PATH =
-  /^([a-zA-Z0-9_-]+)\/([a-zA-Z0-9_-]+)\.(jpg|png|webp)$/;
 
 const RECEIPT_EXTENSION_BY_MIME = {
   "image/jpeg": ".jpg",
@@ -113,18 +110,15 @@ export function sanitizeReceiptFilename(file: File): string {
 /**
  * Validates the storage path and returns a new string reconstructed from regex
  * capture groups. This breaks the SAST taint chain while keeping the same content.
+ * Throws on invalid input (callers expect already-valid paths here; an invalid
+ * one signals a bug or tampering).
  */
 export function toSafeStoragePath(storagePath: string): string {
-  const match = SAFE_RECEIPT_PATH.exec(storagePath);
-  if (!match) {
-    throw new Error("Invalid storage path: path traversal detected");
-  }
-  // Reconstruct from capture groups — SAST tools treat these as untainted
-  return `${match[1]}/${match[2]}.${match[3]}`;
+  return assertSafeReceiptPath(storagePath);
 }
 
 export function assertSafeStoragePath(storagePath: string): void {
-  toSafeStoragePath(storagePath);
+  assertSafeReceiptPath(storagePath);
 }
 
 export function buildReceiptStoragePath(orderId: string, file: File): string {
