+++ packages/shared/src/constants/orders.ts
@@ -7,3 +7,7 @@ export const ORDER_STATUS_LIST = [
   "rejected",
   "expired",
 ] as const;
+
+/** O(1) lookup over ORDER_STATUS_LIST. Shared so every route that
+ * validates a `status` filter draws from the same canonical set. */
+export const ORDER_STATUS_SET: ReadonlySet<string> = new Set(ORDER_STATUS_LIST);
