+++ packages/shared/src/constants/receipts.ts
@@ -1,0 +1,5 @@
+/** Supabase Storage bucket name where uploaded payment receipts live. */
+export const RECEIPTS_BUCKET = "receipts";
+
+/** TTL (seconds) for signed receipt URLs returned to the browser. */
+export const RECEIPT_URL_TTL_SECONDS = 60 * 60;
